apps/backend/requirements.txt
fastapi
uvicorn
chromadb
openai
beautifulsoup4
pydantic
httpx
python-dotenv 